Type TS language features extension API
