+++ src/index.ts
@@ -1,4 +1,5 @@
 import * as vscode from "vscode";
+import { getTypeScriptLanguageFeaturesExtensionAPI } from "./vscode-ts-extension";
 
 const pluginId = "typescript-sql-tagged-template-plugin";
 const configurationSection = "sql-tagged-template-literals";
@@ -7,22 +8,6 @@ interface Configuration {
   schemaFile?: string;
   defaultSchemaName?: string;
 }
-
-const getTypeScriptLanguageFeaturesExtensionAPI = async () => {
-  const extension = vscode.extensions.getExtension(
-    "vscode.typescript-language-features"
-  );
-  if (!extension) {
-    return;
-  }
-
-  await extension.activate();
-  if (!extension.exports || !extension.exports.getAPI) {
-    return;
-  }
-
-  return extension.exports.getAPI(0);
-};
 
 const getConfiguration = (): Configuration => {
   const config = vscode.workspace.getConfiguration(configurationSection);
@@ -39,9 +24,6 @@ const getConfiguration = (): Configuration => {
   };
 };
 
-const synchronizeConfiguration = (api: any) =>
-  api.configurePlugin(pluginId, getConfiguration());
-
 export async function activate(context: vscode.ExtensionContext) {
   const api = await getTypeScriptLanguageFeaturesExtensionAPI();
   if (!api) {
@@ -51,12 +33,12 @@ export async function activate(context: vscode.ExtensionContext) {
   vscode.workspace.onDidChangeConfiguration(
     e => {
       if (e.affectsConfiguration(configurationSection)) {
-        synchronizeConfiguration(api);
+        api.configurePlugin(pluginId, getConfiguration());
       }
     },
     undefined,
     context.subscriptions
   );
 
-  synchronizeConfiguration(api);
+  api.configurePlugin(pluginId, getConfiguration());
 }
